QNM Suite

Einstein.QNMSuite.contfrac_lentz

Einstein.QNMSuite.qnm_kerr_cf

Einstein.QNMSuite.qnm_kerr_cheb

Einstein.QNMSuite.qnm_kerr_cheb_δ!

Einstein.QNMSuite.qnm_kerr_di

Einstein.QNMSuite.qnm_kerr_radial_cf

Einstein.QNMSuite.qnm_kerr_radial_di_horizon_series

Einstein.QNMSuite.qnm_kerr_radial_di_inf_series

Einstein.QNMSuite.qnm_kerr_radial_di_rhs

Einstein.QNMSuite.qnm_kerr_radial_di_δ!

Einstein.QNMSuite.qnm_polyeig

Einstein.QNMSuite.qnm_schw_cheb

Einstein.QNMSuite.qnm_schw_reggepoleexp

Einstein.QNMSuite.contfrac_lentz — Method

contfrac_lentz([T=Float64], a::Function, b::Function, tol::T, min_iter::Integer, max_iter::Integer) where {T<:AbstractFloat}

Compute the continued fraction

\[f(x)=b_0+\frac{a_1}{b_1+\frac{a_2}{b_2+\frac{a_3}{b_3+\frac{a_4}{b_4+\frac{a_5}{b_5+\cdots}}}}}\]

using modified Lentz's method. Translated from duetosymmetry/qnm.

Arguments

a: A function that returns the aᵢ terms.

b: A function that returns the bᵢ terms.

tol: The tolerance for convergence.

min_iter: The minimum number of iterations to perform.

max_iter: The maximum number of iterations to perform.

Returns

fᵢ: The value of the continued fraction.

errorᵢ: The estimated error.

i: The number of iterations performed.

Examples

Compute the square root of two using continued fractions

\[\sqrt{2} = 1 + \frac{1}{2 + \frac{1}{2 + \frac{1}{2 + \frac{1}{2 + \cdots}}}} \approx 1.414213562373095\]

a(i) = 1
b(i) = i == 0 ? 1 : 2
contfrac_lentz(Float64, a, b, 10*eps(Float64), 50, 1000)

Compute Golden Ratio

\[\phi = 1 + \frac{1}{1 + \frac{1}{1 + \frac{1}{1 + \cdots}}} \approx 1.618033988749895\]

a(i) = 1
b(i) = 1
contfrac_lentz(Float64, a, b, 10*eps(Float64), 50, 1000)

References

qnm/qnm/contfrac.py at master · duetosymmetry/qnm

[Pre07, Ste19, Len76, TB86, Tem24]

Einstein.QNMSuite.qnm_kerr_cf — Method

qnm_kerr_cf(params::QNMKerrCFParams{TR}; alg=RobustMultiNewton(autodiff=AutoFiniteDiff()), kwargs...)

Find the Kerr QNM using the Leaver's method for the radial equation and the Cook-Zalutskiy approach for the angular sector.

Arguments

params: QNMKerrCFParams object containing the Kerr parameters and initial guess

alg: Nonlinear algorithm to use for the eigenvalue search (default: RobustMultiNewton with AutoFiniteDiff)

kwargs: Additional keyword arguments to pass to the nonlinear solver

Einstein.QNMSuite.qnm_kerr_cheb — Method

qnm_kerr_cheb(params::QNMKerrChebParams{TR}; alg=RobustMultiNewton(autodiff=AutoFiniteDiff()), kwargs...)

Find the Kerr QNM using the Ultraspherical spectral method for the radial equation and the Cook-Zalutskiy approach for the angular sector.

Arguments

params: QNMKerrChebParams object containing the Kerr parameters and initial guess

alg: Nonlinear algorithm to use for the eigenvalue search (default: RobustMultiNewton with AutoFiniteDiff)

kwargs: Additional keyword arguments to pass to the nonlinear solver

Einstein.QNMSuite.qnm_kerr_cheb_δ! — Method

qnm_kerr_cheb_δ!(cache, ω, l; l_max=20)

Perform one iteration step in the QNM eigenvalue search.

Arguments

cache: Pre-computed QNMKerrChebCache object

ω: Current frequency guess

l: Angular mode number

l_max: Maximum l value for angular eigenvalue calculation (default: 20)

Returns

Difference between the seperation constant between the radial and angular equations

Einstein.QNMSuite.qnm_kerr_di — Method

qnm_kerr_di(params::QNMKerrDIParams{TR}, alg::AbstractNonlinearAlgorithm=RobustMultiNewton(autodiff=AutoFiniteDiff()); kwargs...) where {TR<:AbstractFloat}

Find the Kerr QNM using the direct integration method for the radial equation and the Cook-Zalutskiy approach for the angular sector.

Arguments

params: QNMKerrDIParams object containing the Kerr parameters and initial guess

alg: Nonlinear algorithm to use for the eigenvalue search (default: RobustMultiNewton with AutoFiniteDiff)

kwargs: Additional keyword arguments to pass to the nonlinear solver

Einstein.QNMSuite.qnm_kerr_radial_cf — Method

qnm_kerr_radial_cf(
    ::Type{TR},
    a::TR,
    s::Integer,
    m::Integer,
    A::Complex{TR},
    ω::Complex{TR};
    n_inv::Integer=0,
    cf_tol::TR=typetol(TR),
    cf_N_min::Integer=300,
    cf_N_max::Integer=100000,
) where {TR<:AbstractFloat}

Calculate the radial function using the Leaver scheme [CZ14].

Arguments

TR: Type of the floating-point number.

a::TR: Black hole spin.

s::Integer: spin weight of the field.

m::Integer: Azimuthal number.

A::Complex{TR}: Angular separation constant.

ω::Complex{TR}: QNM frequency.

n_inv::Integer: Number of inversions.

cf_tol::TR: Tolerance for convergence.

cfNmin::Integer: Minimum number of iterations.

cfNmax::Integer: Maximum number of iterations.

Returns

inv_cf::Complex{TR}: The inverse continued fraction.

error::TR: The error.

iter::Integer: The number of iterations.

Leaver scheme

\[\begin{aligned} & D_0=\delta=1+s+2 \xi \\ & D_1=4 p-2 \alpha+\gamma-\delta-2 \\ & D_2=2 \alpha-\gamma+2 \\ & D_3=\alpha(4 p-\delta)-\sigma, \\ & D_4=\alpha(\alpha-\gamma+1) \end{aligned}\]

\[\begin{aligned} \alpha_n & \equiv n^2+\left(D_0+1\right) n+D_0 \\ \beta_n & \equiv-2 n^2+\left(D_1+2\right) n+D_3 \\ \gamma_n & \equiv n^2+\left(D_2-3\right) n+D_4-D_2+2 \end{aligned}\]

The general continued fraction:

\[0=\beta_0-\frac{\alpha_0 \gamma_1}{\beta_1-} \frac{\alpha_1 \gamma_2}{\beta_2-} \frac{\alpha_2 \gamma_3}{\beta_3-} \ldots\]

The truncated version of the continued fraction:

\[\operatorname{Cf}(\mathrm{N}) \equiv \beta_0-\frac{\alpha_0 \gamma_1}{\beta_1-} \frac{\alpha_1 \gamma_2}{\beta_2-} \frac{\alpha_2 \gamma_3}{\beta_3-} \ldots \frac{\alpha_{\mathrm{N}-1} \gamma_{\mathrm{N}}}{\beta_{\mathrm{N}}+\alpha_{\mathrm{N}} \mathrm{r}_{\mathrm{N}}}\]

The nth inversion of this truncated continued fraction:

\[\begin{aligned} \operatorname{Cf}(\mathrm{n} ; \mathrm{N}) \equiv & \beta_n-\frac{\alpha_{n-1} \gamma_n}{\beta_{n-1}-} \frac{\alpha_{n-2} \gamma_{n-1}}{\beta_{n-2}-} \ldots \frac{\alpha_0 \gamma_1}{\beta_0} \\ & -\frac{\alpha_n \gamma_{n+1}}{\beta_{n+1}-} \frac{\alpha_{n+1} \gamma_{n+2}}{\beta_{n+2}-} \cdots \frac{\alpha_{N-1} \gamma_N}{\beta_N+\alpha_N r_N} \end{aligned}\]

with $\operatorname{Cf}(0 ; \mathrm{N}) \equiv \operatorname{Cf}(\mathrm{N})$ and $0 \leq n < N$

References

Cook and Zalutskiy [CZ14]

qnm/qnm/radial.py at master · duetosymmetry/qnm

Einstein.QNMSuite.qnm_kerr_radial_di_horizon_series — Method

qnm_kerr_radial_di_horizon_series(
    s::Integer,
    ρ::TR,
    a::TR,
    m::Integer,
    ω::Complex{TR},
    Λ::Complex{TR},
    series_order::Integer,
) where {TR<:AbstractFloat}

Calculate the series expansion of the radial equation at the horizon.

Arguments

s::Integer: spin weight of the field.

ρ::TR: compactified radial coordinate.

a::TR: black hole spin parameter.

m::Integer: azimuthal mode number.

ω::Complex{TR}: QNM frequency.

Λ::Complex{TR}: separation constant.

series_order::Integer: order of the series expansion.

Returns

R0::Complex{TR}: value of the radial function.

dR0::Complex{TR}: value of the radial derivative.

References

Ripley [Rip22]

Einstein.QNMSuite.qnm_kerr_radial_di_inf_series — Method

qnm_kerr_radial_di_inf_series(
    s::Integer,
    ρ::TR,
    a::TR,
    m::Integer,
    ω::Complex{TR},
    Λ::Complex{TR},
    series_order::Integer,
) where {TR<:AbstractFloat}

Calculate the series expansion of the radial equation at the horizon.

Arguments

s::Integer: spin weight of the field.

ρ::TR: compactified radial coordinate.

a::TR: black hole spin parameter.

m::Integer: azimuthal mode number.

ω::Complex{TR}: QNM frequency.

Λ::Complex{TR}: separation constant.

series_order::Integer: order of the series expansion.

Returns

R0::Complex{TR}: value of the radial function.

dR0::Complex{TR}: value of the radial derivative.

References

Ripley [Rip22]

Einstein.QNMSuite.qnm_kerr_radial_di_rhs — Method

qnm_kerr_radial_di_rhs(
    u::SVector{2,Complex{TR}},
    cache::QNMKerrDICache{TR},
    ρ::TR,
) where {TR<:AbstractFloat}

Calculate the right-hand side of the radial equation in hyperboloidal coordinates.

Arguments

u::SVector{2,Complex{TR}}: radial function and its derivative.

cache::QNMKerrDICache{TR}: cache object.

ρ::TR: compactified radial coordinate.

References

Ripley [Rip22]

Einstein.QNMSuite.qnm_kerr_radial_di_δ! — Method

qnm_kerr_radial_di_δ!(
    x::SVector{2,TR},
    cache::QNMKerrDICache{TR},
) where {TR<:AbstractFloat}

Integrate the radial equation from infinity (ρmin) to the horizon (ρmax) and calculate the difference compared to

series expansion at the horizon (Natural boundary condition)

0 (Dirichlet boundary condition)

Arguments

x::SVector{2,TR}: real and imaginary parts of the QNM frequency.

cache::QNMKerrDICache{TR}: cache object.

Returns

δ::SVector{2,TR}: real and imaginary parts of the difference.

References

Ripley [Rip22]

Einstein.QNMSuite.qnm_polyeig — Method

A, E = qnm_polyeig(TFC, pep::AbstractVector{<:AbstractMatrix}) where TFC <: Union{AbstractFloat, Complex{<:AbstractFloat}}

Linearizes a polynomial eigenvalue problem (PEP) a to the companion form, , as in the paper by Mehrmann and Voss. More precisely, for a k-th degree PEP with n-by-n coefficient matrices, this returns matrices A and E, both kn-by-kn, corresponding to the linearized problem

\[Ax = λEx\]

References

Mehrmann and Voss [MV04]

Einstein.QNMSuite.qnm_schw_cheb — Method

qnm_schw_cheb(::Type{TR}, s::Integer, ℓ::Integer, n::Integer, potential::SchwPType.T; σ_min::TR=zero(TR), σ_max::TR=one(TR), lo_bc::BCType.T=Natural, hi_bc::BCType.T=Natural)

Construct a polynomial eigenvalue problem for the Schwarzschild spacetime using the hyperboloidal coordinates and the ultraspherical spectral method. The coordinate transformation from standard Schwarzschild coordinates to hyperboloidal coordinates is given by (we use $M = 1$ in the code):

\[\begin{align} \sigma &= \frac{2 M}{r} \\ \tau &= t + 2 M \left( \ln\sigma + \ln(1 - \sigma) - \frac{1}{\sigma} \right) \\ r_* &= 2 M \left(\frac{1}{\sigma} + \ln(1 - \sigma) - \ln\sigma \right) \end{align}\]

Arguments

s::Integer: Spin

ℓ::Integer: Angular number

n::Integer: Number of Chebyshev points

potential::SchwPType.T: Potential type

σ_min::TR=zero(TR): Minimum value of the radial coordinate (hyperboloidal slicing)

σ_max::TR=one(TR): Maximum value of the radial coordinate (hyperboloidal slicing)

lo_bc::BCType.T=BCType.Natural: Boundary condition at the lower boundary, either Natural or Dirichlet.

hi_bc::BCType.T=BCType.Natural: Boundary condition at the upper boundary, either Natural or Dirichlet.

Returns

A array of matrices representing the Polynomial eigenvalue problem, which can be solved using qnm_polyeig or solvers from the NonlinearEigenproblems.jl package, such as polyeig.

References

Jaramillo, Panosso Macedo and Al Sheikh [JPA21] and Panosso Macedo [Pan24]

Einstein.QNMSuite.qnm_schw_reggepoleexp — Method

qnm_schw_reggepoleexp(::Type{TR}, s::TI, l::TI, n::TI) where {TR<:AbstractFloat,TI<:Integer}

Compute the high $\ell$ asymptotic expansion of Schwarzschild QNM frequency, using the method of Dolan and Ottewill [DO09].

The QNM frequency can be written as an expansion in inverse powers of $L=\ell+\frac{1}{2}$

\[\omega_{l n}=\varpi_{-1}^{(n)} L+\varpi_0^{(n)}+\varpi_1^{(n)} L^{-1}+\varpi_2^{(n)} L^{-2}+\ldots\]

The lowest expansion coefficients for arbitrary spin $\beta=1-s^2$ and arbitrary overtone number $n$ are given by

\[\begin{aligned} & \sqrt{27} \varpi_{-1}^{(n)}=1 \\ & \sqrt{27} \varpi_0^{(n)}=-i N \\ & \sqrt{27} \varpi_1^{(n)}=\frac{\beta}{3}-\frac{5 N^2}{36}-\frac{115}{432} \\ & \sqrt{27} \varpi_2^{(n)}=-i N\left[\frac{\beta}{9}+\frac{235 N^2}{3888}-\frac{1415}{15552}\right] \\ & \sqrt{27} \varpi_3^{(n)}=-\frac{\beta^2}{27}+\frac{204 N^2+211}{3888} \beta+\frac{854160 N^4-1664760 N^2-776939}{40310784} \\ & \sqrt{27} \varpi_4^{(n)}=i N\left[\frac{\beta^2}{27}+\frac{1100 N^2-2719}{46656} \beta+\frac{11273136 N^4-52753800 N^2+66480535}{2902376448}\right] \end{aligned}\]

Arguments

TR: Type of the floating-point number.

s: Spin weight of the field of interest.

l: Multipole number of interest.

n: Overtone number of interest.

References

[DO09]