General Utilities

GRSuite.Utils.dot_kahan

GRSuite.Utils.dot_kahan_neumaier

GRSuite.Utils.dot_xsum

GRSuite.Utils.sum_kahan

GRSuite.Utils.sum_kahan_neumaier

GRSuite.Utils.sum_xsum

GRSuite.Utils.dot_kahan — Method

dot_kahan(v1::StridedVector{T}, v2::StridedVector{T}) where {T<:Number}

Compute the dot product using Kahan summation algorithm to reduce numerical errors.

Note

Slower than dot_xsum for large vectors, but faster for small vectors.

Similar performance to dot_kahan_neumaier

Both input vectors must have the same length, which is not checked for performance reasons.

GRSuite.Utils.dot_kahan_neumaier — Method

dot_kahan_neumaier(v1::StridedVector{T}, v2::StridedVector{T}) where {T<:Number}

Neumaier's variant of Kahan summation algorithm to reduce numerical errors.

Note

Slower than dot_xsum for large vectors, but faster for small vectors.

Similar performance to dot_kahan

Uses loop unrolling for better performance while maintaining Kahan summation's

numerical stability. Processes two elements per iteration when possible.

References

JuliaMath/KahanSummation.jl

GRSuite.Utils.dot_xsum — Method

dot_xsum(x::StridedVector{T}, y::StridedVector{T}) where {T<:Real}

Compute the dot product of two vectors using extended precision accumulation. Uses the xsum package for improved numerical accuracy.

Note

Very fast for large vectors, but a bit slower than Kahan summation for small vectors.

Both input vectors must have the same length, which is not checked for performance reasons.

References

Neal [Nea15]

JuliaMath/Xsum.jl

Radford Neal / xsum · GitLab

GRSuite.Utils.sum_kahan — Method

sum_kahan(v::StridedVector{T}) where {T<:Number}

Compute the sum using Kahan summation algorithm to reduce numerical errors.

Note

Slower than sum_xsum for large vectors, but faster for small vectors.

Similar performance to sum_kahan_neumaier

GRSuite.Utils.sum_kahan_neumaier — Method

sum_kahan_neumaier(v::StridedVector{T}) where {T<:Number}

Neumaier's variant of Kahan summation algorithm to reduce numerical errors.

Note

Slower than sum_xsum for large vectors, but faster for small vectors.

Similar performance to sum_kahan

Uses loop unrolling for better performance while maintaining Kahan summation's

numerical stability. Processes two elements per iteration when possible.

References

JuliaMath/KahanSummation.jl

GRSuite.Utils.sum_xsum — Method

sum_xsum(vec::StridedVector{T}) where {T<:Union{Float32,Float64}}

Compute the sum of a vector using extended precision accumulation. Uses the xsum package for improved numerical accuracy.

Note

Very fast for large vectors, but a bit slower than Kahan summation for small vectors (n ⪅ 80)

References

Neal [Nea15]

JuliaMath/Xsum.jl

Radford Neal / xsum · GitLab